Se cambian las pantallas de Altam baja y modificacion para que no se repita la funcionalidad

+++ TallerMecanico/src/servicios/AutoBusinessObjectImpl.java
@@ -5,6 +5,7 @@ import java.util.List;
 import dao.AutoDAO;
 import dao.AutoDAOImpl;
 import entidades.Auto;
+import excepciones.CamposVaciosException;
 import excepciones.ExistingCarException;
 import excepciones.NoIdObtainedException;
 import excepciones.NonExistingCarException;
@@ -20,13 +21,25 @@ public class AutoBusinessObjectImpl implements AutoBusinessObject {
 	}
 
 	@Override
-	public Auto selectAuto(String patente) throws TallerMecanicoException, NonExistingCarException {
-		return autoDAO.selectAuto(patente);
+	public Auto selectAuto(String patente)
+			throws TallerMecanicoException, NonExistingCarException, CamposVaciosException {
+		if (!patente.equals("")) {
+			return autoDAO.selectAuto(patente);
+		} else {
+			throw new CamposVaciosException("No puede haber espacios sin completar");
+		}
+
 	}
 
 	@Override
-	public void insertAuto(Auto auto) throws TallerMecanicoException, ExistingCarException, NoIdObtainedException {
-		autoDAO.insertAuto(auto);
+	public void insertAuto(Auto auto)
+			throws TallerMecanicoException, ExistingCarException, NoIdObtainedException, CamposVaciosException {
+		if (auto.getPatente() != null && auto.getMarca() != null && auto.getModelo() != null && auto.getColor() != null
+				&& auto.getCantidadPuertas() != 0 && auto.getAño() != null && auto.getKilometraje() != 0) {
+			autoDAO.insertAuto(auto);
+		} else {
+			throw new CamposVaciosException("No puede haber espacios sin completar");
+		}
 	}
 
 	@Override
@@ -35,8 +48,13 @@ public class AutoBusinessObjectImpl implements AutoBusinessObject {
 	}
 
 	@Override
-	public void deleteAuto(String patente) throws TallerMecanicoException, NonExistingCarException {
-		autoDAO.deleteAuto(patente);
+	public void deleteAuto(String patente)
+			throws TallerMecanicoException, NonExistingCarException, CamposVaciosException {
+		if (!patente.equals("")) {
+			autoDAO.deleteAuto(patente);
+		} else {
+			throw new CamposVaciosException("No puede haber espacios sin completar");
+		}
 	}
 
 	@Override
